@@ -1,14 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/throw';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/delay';
-import { IConcepto } from '../models/concepto';
-import { UsersService } from './users.service';
+import { Observable } from 'rxjs';
 import { UrlService } from './url.service';
+import { tap } from 'rxjs/operators';
 
 @Injectable()
 export class ConceptoService {
@@ -20,7 +14,7 @@ export class ConceptoService {
     getConceptos(): Observable<any[]> {
         return this._http.get<any[]>(this._urlService.urlGetConceptos())
                         //.delay(3000)
-                        .do(data => JSON.stringify(data));
+                        .pipe(tap(data => JSON.stringify(data)));
     }
 
     insertConcepto(descripcion:string, credito:boolean) : Observable<void> {
