@@ -1,11 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/throw';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/delay';
+import { Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
 
 @Injectable()
 export class IpService {
@@ -17,7 +13,7 @@ export class IpService {
     let url = "https://ipapi.co/json/";
     
     return this._http.get(url)
-                    .do(data => JSON.stringify(data))
+                    .pipe(tap(data => JSON.stringify(data)))
   }
 
   
