@@ -9,5 +9,5 @@ import 'hammerjs';
 export class AppComponent {
   pageTitle = 'app';
 
-  constructor() {  }   
+  constructor() { }   
 }
