@@ -1,6 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AppComponent } from './app.component';
 import { DiarioComponent } from './diario/diario.component';
 import { RouterModule, Routes, PreloadAllModules } from '@angular/router';
@@ -8,20 +8,12 @@ import { MensualComponent } from './mensual/mensual.component';
 import { AnualComponent } from './anual/anual.component';
 import { HistoricoComponent } from './historico/historico.component';
 import { ConceptosComponent } from './conceptos/conceptos.component';
-import { ButtonsModule } from 'ngx-bootstrap/buttons';
-import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
-import { AccordionModule } from 'ngx-bootstrap/accordion';
-import { BsDatepickerModule } from 'ngx-bootstrap/datepicker';
-import { SumarydiaComponent } from './sumarydia/sumarydia.component';
 import { SumarymesComponent } from './sumarymes/sumarymes.component';
 import { HttpClientModule } from '@angular/common/http';
 import { SumaryMonthService } from './services/sumary-month.service';
 import { SumaryAnioService } from './services/sumary-anio.service';
 import { SumaryHistoricoService } from './services/sumary-historico.service';
-import { BsLocaleService } from 'ngx-bootstrap/datepicker';
 import { DiarioService } from './services/diario.service';
-import { ModalModule, BsModalService } from 'ngx-bootstrap/modal';
-import { CurrencyMaskModule } from "ng2-currency-mask";
 import { LoginModule } from './login/login.module';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { LoginComponent } from './login/login.component';
@@ -29,8 +21,6 @@ import { AuthGuard } from './guards/auth.guard';
 import { UsersService } from './services/users.service';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { AuthInterceptor } from './interceptors/AuthInterceptor';
-import { APP_CONFIG, APP_DI_CONFIG } from "./app.config/app-config.constants";
-import { LocalizacionService } from './services/localizacion.service';
 import { LOCALE_ID } from '@angular/core';
 import { ConceptoService } from './services/concepto.service';
 import { RegistracionComponent } from './registracion/registracion.component';
@@ -41,8 +31,46 @@ import { ErrorMessageComponent } from './error-message/error-message.component';
 import { SumaryanioComponent } from './sumaryanio/sumaryanio.component';
 import { UserprofileComponent } from './userprofile/userprofile.component';
 import { SumaryhistoricoComponent } from './sumaryhistorico/sumaryhistorico.component';
-import { HashLocationStrategy, LocationStrategy, CurrencyPipe } from '@angular/common';
+import { HashLocationStrategy, LocationStrategy, CurrencyPipe, registerLocaleData, DatePipe } from '@angular/common';
 import { IpService } from './services/ip.service';
+import { MatToolbarModule, 
+  MatCardModule, 
+  MatButtonModule, 
+  MatButtonToggleModule, 
+  MatBadgeModule, 
+  MatProgressSpinnerModule,
+  MatFormFieldModule,
+  MatNativeDateModule,
+  MatInputModule,
+  MatSelectModule,
+  MatSlideToggleModule,
+  MatSidenavModule,
+  MatExpansionModule,
+  MatGridListModule,
+  MatTabsModule,
+  MatDialogModule,
+  MatSnackBarModule,
+  MatTableModule,
+  MatDatepickerModule, 
+  MatTooltipModule} from  '@angular/material';
+import { MatIconModule  } from '@angular/material/icon';
+import { MatChipsModule } from '@angular/material/chips';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatListModule } from '@angular/material/list';
+import { AboutComponent } from './about/about.component';
+import { DiarioEnterComponent } from './diario-enter/diario-enter.component';
+import { TwoDigitDecimaNumberDirective } from './directives/TwoDigitDecimaNumberDirective';
+import { FormatingService } from './sharedServices/formatingService';
+import localeEs from '@angular/common/locales/es';
+import localeEn from '@angular/common/locales/en';
+import { SaldoComponent } from './saldo/saldo.component';
+import { SaldoAbiertoComponent } from './saldo-abierto/saldo-abierto.component';
+import { SaldoItemComponent } from './saldo-item/saldo-item.component';
+import { WelcomeComponent } from './welcome/welcome.component';
+import { ConceptoDialogComponent } from './concepto-dialog/concepto-dialog.component';
+import { CalculationService } from './sharedServices/calculationService';
+registerLocaleData(localeEs, localeEs); 
+registerLocaleData(localeEn, localeEn); 
 
 const routes: Routes = [    
   { path: '', component: LoginComponent},
@@ -68,7 +96,6 @@ const routes: Routes = [
     AnualComponent,
     HistoricoComponent,
     ConceptosComponent,
-    SumarydiaComponent,
     SumarymesComponent,
     DashboardComponent,
     RegistracionComponent,
@@ -76,47 +103,68 @@ const routes: Routes = [
     ErrorMessageComponent,
     SumaryanioComponent,
     UserprofileComponent,
-    SumaryhistoricoComponent
+    SumaryhistoricoComponent,
+    AboutComponent,
+    DiarioEnterComponent,    
+    TwoDigitDecimaNumberDirective, SaldoComponent, SaldoAbiertoComponent, SaldoItemComponent, WelcomeComponent, ConceptoDialogComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
+    ReactiveFormsModule,
     LoginModule,
-    ModalModule,
     HttpClientModule,
-    CurrencyMaskModule,
-    ButtonsModule.forRoot(),
-    BsDatepickerModule.forRoot(),
-    BsDropdownModule.forRoot(),
-    AccordionModule.forRoot(),
+    MatToolbarModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
+    MatButtonToggleModule,
+    MatBadgeModule,
+    MatFormFieldModule,
+    MatProgressSpinnerModule,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatInputModule,
+    MatSelectModule,
+    MatSlideToggleModule,
+    MatSidenavModule,
+    MatExpansionModule,
+    MatGridListModule,
+    MatTabsModule,
+    MatDialogModule,
+    MatSnackBarModule,
+    MatTableModule,
+    MatListModule,
+    MatRadioModule,
+    MatChipsModule,
+    MatTooltipModule,
     RouterModule.forRoot(routes, { useHash: false, preloadingStrategy: PreloadAllModules }),
   ],
   providers: [SumaryMonthService,
               SumaryAnioService,
               SumaryHistoricoService,
               DiarioService, 
-              BsModalService,
               UsersService,
               ConceptoService, 
               IpService,
               HelperService,
+              FormatingService,
+              CalculationService,
               UrlService,
               CurrencyPipe,
-              AuthGuard ,
+              DatePipe,
+              AuthGuard,
               {
                 provide: HTTP_INTERCEPTORS,
                 useClass: AuthInterceptor,
                 multi: true,
               },
-              {
-                provide: APP_CONFIG,
-                useValue: APP_DI_CONFIG
-              },
               {provide: LocationStrategy, useClass: HashLocationStrategy},
-              LocalizacionService,
               { provide: LOCALE_ID,
-                useValue: window.navigator.language }
+                useValue: window.navigator.language.split('-')[0]
+              }              
             ],
+  entryComponents: [ AboutComponent, DiarioEnterComponent, SaldoAbiertoComponent, WelcomeComponent, ConceptoDialogComponent ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
