@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ConceptoService } from '../services/concepto.service';
 import { IConcepto } from '../models/concepto';
-import { NgForm } from '@angular/forms';
 import { HelperService } from '../services/helper.service';
+import { MatSnackBar, MatDialog } from '@angular/material';
+import { ConceptoDialogComponent } from '../concepto-dialog/concepto-dialog.component';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-conceptos',
@@ -10,101 +12,58 @@ import { HelperService } from '../services/helper.service';
   styleUrls: ['./conceptos.component.css']
 })
 export class ConceptosComponent implements OnInit {
-  conceptos: any[];
-  errorMessageGridConceptos: string = "";
-  model: IConcepto = new IConcepto();
-  pantallaActual: number; /* 0-ninguna / 1 - Alta / 2 - Modificacion */
-  pantallaTitulo: string;
-  operationMessage: string = "";
-  operationMessageStatus: number = 0; /* 0 - OK / 1 - Error */
+  displayedColumns: string[] = ['concepto', 'tipo'];
   loading: boolean = false;
-  loadingGridConceptos: boolean = false;
+  conceptos: IConcepto[] = [];
+  private getConceptosSubscription: Subscription;
+  private conceptoDialogSubscription: Subscription;
 
-  constructor(
-          private _conceptoService: ConceptoService,
-          private _helperService: HelperService) {
-    this.errorMessageGridConceptos = "";
-    this.model.descripcion = "";
-    this.model.id = "";
-    this.model.suma = false;
-    this.pantallaActual = 0;
+  constructor(private _conceptoService: ConceptoService,
+              private _helperService: HelperService,
+              public conceptoDialog: MatDialog,
+              public snackBar: MatSnackBar){  }
+
+  ngOnInit() {
     this.getConceptos();
   }
 
-  ngOnInit() {
+  ngOnDestroy(): void {
+    this.unsubscribeGetConceptos();
+    this.unsubscribeConceptoDialog();
   }
 
-  changeCredDeb(value: boolean) {
-    this.model.suma = value;
+  unsubscribeGetConceptos(): void {
+    if (this.getConceptosSubscription){ this.getConceptosSubscription.unsubscribe(); }    
+  }
+
+  unsubscribeConceptoDialog(): void {
+    if (this.conceptoDialogSubscription){ this.getConceptosSubscription.unsubscribe(); }    
   }
 
   getConceptos() {
-    this.loadingGridConceptos = true;
-    this.errorMessageGridConceptos = "";
-    this._conceptoService.getConceptos()
+    this.loading = true;
+    this.unsubscribeGetConceptos();
+    this.getConceptosSubscription = this._conceptoService.getConceptos()
         .subscribe(
             data => { 
               this.conceptos = data;              
-              this.loadingGridConceptos = false;
+              this.loading = false;
             },
             error => {
-              this.loadingGridConceptos = false; 
-              this.errorMessageGridConceptos = this._helperService.getErrorMessage(error);
+              this.loading = false; 
+              this.snackBar.open(this._helperService.getErrorMessage(error), '', { duration: 2000, panelClass: ['error-snackbar'], direction: 'ltr', verticalPosition: 'bottom' });
             });
   }
 
-  cambiarPantalla(pantallaNueva: number, conceptoSeleccionado: any) {
-    this.operationMessage = "";
-    this.pantallaActual = pantallaNueva;
-    this.model.descripcion = "";
-    this.pantallaTitulo = (pantallaNueva == 1) ? "Alta de Concepto" : "Modificacion de Concepto";
-    if (conceptoSeleccionado != null) {
-      this.model.descripcion = conceptoSeleccionado.descripcion;
-      this.model.suma = conceptoSeleccionado.credito;
-      this.model.id = conceptoSeleccionado._id;
-    } 
+  openConceptoDialog(concepto: IConcepto){    
+    let dialogRef = this.conceptoDialog.open(ConceptoDialogComponent, { data: {concepto} });    
+    
+    this.unsubscribeConceptoDialog();
+    this.getConceptosSubscription = dialogRef.afterClosed().subscribe(result => {
+      if (result !== undefined){
+        this.getConceptos();
+      }
+    });
   }
-
-  aceptar(f: NgForm): void {
-    this.loading = true;
-    this.model.descripcion = this.model.descripcion[0].toUpperCase() + this.model.descripcion.slice(1);
-
-    // ALTA
-    if (this.pantallaActual == 1) {
-      this._conceptoService.insertConcepto(
-                                      this.model.descripcion, 
-                                      this.model.suma)
-                  .subscribe(
-                    data => {
-                      f.resetForm();
-                      this.operationMessage = "Carga Exitosa!";
-                      this.operationMessageStatus = 0;
-                      this.loading = false;
-                      this.getConceptos();
-                    },
-                    error => {
-                      this.operationMessage = this._helperService.getErrorMessage(error);
-                      this.operationMessageStatus = 1;
-                      this.loading = false;
-                    });
-    }
-    else {
-      // MODIFICACION      
-      this._conceptoService.updateConcepto(
-                            this.model.id,
-                            this.model.descripcion, 
-                            this.model.suma)
-                  .subscribe(
-                    data => {
-                      this.loading = false;
-                      this.pantallaActual = 0; // cierro el form
-                      this.getConceptos();
-                    },
-                    error => {
-                      this.operationMessage = this._helperService.getErrorMessage(error);
-                      this.operationMessageStatus = 1;
-                      this.loading = false;
-                    });
-    }
-  }
+  
 }
