@@ -1,21 +1,27 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
 import { UsersService } from '../services/users.service';
 import { IpService } from '../services/ip.service';
+import { Validators, FormGroup, FormBuilder } from '@angular/forms';
+import { MatSnackBar } from '@angular/material';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.css']
 })
-export class LoginComponent implements OnInit {
-    model: any = {};
+export class LoginComponent implements OnInit, OnDestroy {
     loading = false;
-    error: string = "";
     location: any = {};
-  
+    loginForm: FormGroup;
+    private ipServiceSubscription: Subscription;
+    private loginSubscription: Subscription;
+    
     constructor(private router: Router, 
+                private formBuilder: FormBuilder,
                 private usersService: UsersService,
+                public snackBar: MatSnackBar,
                 private _ipService: IpService) { 
       if (this.usersService.isSessionExpired() === false) {
         this.ingresarApp();
@@ -29,7 +35,7 @@ export class LoginComponent implements OnInit {
       this.location.platform = window.navigator.platform;
       this.location.userAgent = window.navigator.userAgent;
 
-      this._ipService.getClientIp()
+      this.ipServiceSubscription = this._ipService.getClientIp()
         .subscribe(
           data => { 
             this.location.ip = data.ip;
@@ -55,14 +61,31 @@ export class LoginComponent implements OnInit {
       });
     }
 
-    ngOnInit() {
+    ngOnInit() {    
+      this.loginForm = this.formBuilder.group({
+        emailFormControl: ['', [Validators.required, Validators.email]],
+        pwdFormControl: ['', [Validators.required]]
+      });  
+    }
+
+    ngOnDestroy(): void {
+      this.unsubscribeIpService();
+      this.unsubscribeLogin();
+    }
+
+    unsubscribeLogin(): void {
+      if (this.loginSubscription){ this.loginSubscription.unsubscribe(); }    
+    }
+
+    unsubscribeIpService(): void {
+      if (this.loginSubscription){ this.loginSubscription.unsubscribe(); }    
     }
 
     login() {
-      this.error = "";
       this.loading = true;
 
-      this.usersService.login(this.model.username, this.model.password, JSON.stringify(this.location))
+      this.unsubscribeLogin();
+      this.loginSubscription = this.usersService.login(this.loginForm.value.emailFormControl, this.loginForm.value.pwdFormControl, JSON.stringify(this.location))
               .subscribe(
                   data => {
                     if (data === true) {
@@ -70,22 +93,25 @@ export class LoginComponent implements OnInit {
                       this.ingresarApp();          
                     }
                     else {
-                      this.error = 'Acceso Denegado';
+                      this.snackBar.open('Acceso Denegado', '', { duration: 2000, panelClass: ['error-snackbar'], direction: 'ltr', verticalPosition: 'bottom' });
                       this.loading = false;
                     }
                   },
                   error => {
+                    let errorMessage: string;
                     if (error.status === 401) {
-                      this.error = 'Acceso Denegado';
+                      errorMessage = 'Acceso Denegado';
                     }
                     else {
-                      this.error = "Error inesperado";
+                      errorMessage = "Error inesperado: ";
                       if (error.error.errorId != undefined) {
-                        this.error += "<br/>" + error.error.errorId;
+                        errorMessage += error.error.errorId;
                       }
                     }
                     
                     this.loading = false;
+                    this.snackBar.open(errorMessage, '', { duration: 2000, panelClass: ['error-snackbar'], direction: 'ltr', verticalPosition: 'bottom' });
+
                   });
     }
 
