@@ -1,8 +1,87 @@
+import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClientModule } from '@angular/common/http';
+import { MatToolbarModule, 
+         MatCardModule, 
+         MatButtonModule, 
+         MatButtonToggleModule, 
+         MatBadgeModule, 
+         MatProgressSpinnerModule,
+         MatFormFieldModule,
+         MatNativeDateModule,
+         MatInputModule,
+         MatSelectModule,
+         MatSlideToggleModule,
+         MatSidenavModule,
+         MatExpansionModule,
+         MatGridListModule,
+         MatTabsModule,
+         MatDialogModule,
+         MatSnackBarModule,
+         MatTableModule,
+         MatDatepickerModule } from  '@angular/material';
+import { MatIconModule  } from '@angular/material/icon';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import { FormsModule, ReactiveFormsModule }   from '@angular/forms';
 import { LoginComponent } from './login.component';
 import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
-import { FormsModule }   from '@angular/forms';
+import { environment } from 'src/environments/environment';
+
+const routes: Routes = [    
+  { path: 'login', component: LoginComponent}
+];
+
+@NgModule({
+  declarations: [LoginComponent],
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    HttpClientModule,
+    MatToolbarModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
+    MatButtonToggleModule,
+    MatBadgeModule,
+    MatFormFieldModule,
+    MatProgressSpinnerModule,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatInputModule,
+    MatSelectModule,
+    MatSlideToggleModule,
+    MatSidenavModule,
+    MatExpansionModule,
+    MatGridListModule,
+    MatTabsModule,
+    MatDialogModule,
+    MatSnackBarModule,
+    MatTableModule,
+    FormsModule,
+    ReactiveFormsModule,
+    RouterModule.forRoot(routes, { useHash: false, preloadingStrategy: PreloadAllModules }),
+    ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production }),
+  ],
+  providers: []
+})
+export class LoginModule { }
+
+
+
+/*import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { LoginComponent } from './login.component';
+import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
+import { FormsModule, ReactiveFormsModule }   from '@angular/forms';
+import { MatToolbarModule, 
+  MatCardModule, 
+  MatButtonModule, 
+  MatFormFieldModule,
+  MatInputModule } from  '@angular/material';
+import { MatIconModule  } from '@angular/material/icon';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 const routes: Routes = [    
   { path: 'login', component: LoginComponent}
@@ -10,10 +89,20 @@ const routes: Routes = [
 
 @NgModule({
   imports: [
-    CommonModule,
+    BrowserModule,
     FormsModule,
     RouterModule.forRoot(routes, { useHash: false, preloadingStrategy: PreloadAllModules }),
+    BrowserAnimationsModule,
+    MatToolbarModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
+    MatFormFieldModule,
+    MatInputModule,
+    ReactiveFormsModule    
   ],
   declarations: [LoginComponent]
 })
 export class LoginModule { }
+*/
+
