@@ -1,62 +1,97 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { User } from '../models/user';
-import { NgForm } from '@angular/forms';
+import { NgForm, FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { UsersService } from '../services/users.service';
 import { HelperService } from '../services/helper.service';
 import { IpService } from '../services/ip.service';
+import { Router } from '@angular/router';
+import { MatSnackBar, MatDialog, MatDialogRef } from '@angular/material';
+import { WelcomeComponent } from '../welcome/welcome.component';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-registracion',
   templateUrl: './registracion.component.html',
   styleUrls: ['./registracion.component.css']
 })
-export class RegistracionComponent implements OnInit {
-  model: User = new User();
-  operationMessage: string = "";
-  operationMessageStatus: number = 2; /* 0 - OK / 1 - Error */
-  loading: boolean = false;
-  passwordCheck: string;
-  fechaNacim: string;
+export class RegistracionComponent implements OnInit, OnDestroy {
+  registerForm: FormGroup;
+  startDate = new Date((new Date()).getFullYear() - 20, 0, 1);
   monedas = ['$', 'U$D', '€'];
+  loading: boolean = false;
+  dialogRef: MatDialogRef<WelcomeComponent>;
+  private registerSubscription: Subscription;
+  private dialogSubscription: Subscription;
 
-  constructor(private _userService: UsersService,
-              private _helperService: HelperService,
-              private _ipService: IpService) { 
-    this.model.moneda = "$";    
-  }
+  constructor(private formBuilder: FormBuilder,
+              private usersService: UsersService,
+              private helperService: HelperService,
+              private router: Router, 
+              public welcomeDialog: MatDialog,
+              public snackBar: MatSnackBar) { }
 
-  ngOnInit() {
-  }
-
-  aceptar(form: NgForm){
-    this.loading  = true;
-    
-    if (this.passwordCheck != this.model.password) {
-      this.operationMessageStatus = 1;
-      this.operationMessage = "Los Password no coinciden";
-      this.loading  = false;
-      return;
+    ngOnInit() {    
+      this.registerForm = this.formBuilder.group({
+        emailFormControl: ['', [Validators.required, Validators.email]],
+        fechaNacimientoFormControl: ['', [Validators.required]],
+        nameFormControl: ['', [Validators.required]],
+        passwordFormControl: ['', [Validators.required]],
+        passwordRepeatFormControl: [''],
+        monedaFormControl: ['', [Validators.required]]
+      }, {validator: this.checkPasswords });          
     }
 
-    this.model.fechanacimiento = new Date(
-      Number(this.fechaNacim.slice(6,10)),
-      Number(this.fechaNacim.slice(3,5))-1,
-      Number(this.fechaNacim.slice(0,2)),
-      0, 0, 0, 0);
+    ngOnDestroy(): void {
+      this.unsubscribeRegister();
+      this.unsubscribeDialog();
+    }
 
-    this._userService.register(this.model)
+    unsubscribeRegister(): void {
+      if (this.registerSubscription){ this.registerSubscription.unsubscribe(); }    
+    }
+  
+    unsubscribeDialog(): void {
+      if (this.dialogSubscription){ this.dialogSubscription.unsubscribe(); }    
+    }
+
+    checkPasswords(group: FormGroup) {
+      let pass = group.get('passwordFormControl').value;
+      let confirmPass = group.get('passwordRepeatFormControl').value;
+
+      return pass === confirmPass ? null : { notSame: true }     
+    }
+
+    register (){
+      this.loading  = true;
+      let user = this.createUser();      
+      
+      this.unsubscribeRegister();
+      this.registerSubscription = this.usersService.register(user)
           .subscribe(
             data => {
-              this.operationMessageStatus = 0;
-              this.operationMessage = "Alta Exitosa " + this.model.nombre;
-              form.resetForm();
               this.loading  = false;
+              this.dialogRef = this.welcomeDialog.open(WelcomeComponent, { data: {user} });    
+
+              this.unsubscribeDialog();
+              this.dialogSubscription = this.dialogRef.afterClosed().subscribe(() => {
+                this.router.navigate(['/login']);
+              });
             },
             error => {
-              this.operationMessageStatus = 1;  
               this.loading  = false;
-              this.operationMessage = this._helperService.getErrorMessage(error)
+              this.snackBar.open(this.helperService.getErrorMessage(error), '', { duration: 2000, panelClass: ['error-snackbar'], direction: 'ltr', verticalPosition: 'bottom' });              
             });
-  }
+    }
 
-}
+    private createUser() : User{
+      let user = new User();
+      user.email = this.registerForm.value.emailFormControl;
+      user.password = this.registerForm.value.passwordFormControl;
+      user.fechanacimiento = this.registerForm.value.fechaNacimientoFormControl;
+      user.moneda = this.registerForm.value.monedaFormControl;
+      user.nombre = this.registerForm.value.nameFormControl;
+
+      return user;
+    }
+
+  }
